Moving file system to platform
