+++ magnolia/src/core/application.cpp
@@ -2,11 +2,11 @@
 
 #include "core/assert.hpp"
 #include "core/event.hpp"
-#include "core/file_system.hpp"
 #include "core/logger.hpp"
 #include "core/window.hpp"
 #include "physics/physics.hpp"
 #include "platform/file_dialog.hpp"
+#include "platform/file_system.hpp"
 #include "renderer/renderer.hpp"
 #include "renderer/shader.hpp"
 #include "resources/image.hpp"
