+++ magnolia/src/platform/file_system.cpp
@@ -1,4 +1,4 @@
-#include "core/file_system.hpp"
+#include "platform/file_system.hpp"
 
 #include <fstream>
 #include <mutex>
