+++ magnolia/src/platform/file_system.hpp
