+++ magnolia/src/resources/image_loader.cpp
@@ -5,8 +5,8 @@
 #include <set>
 
 #include "core/buffer.hpp"
-#include "core/file_system.hpp"
 #include "core/logger.hpp"
+#include "platform/file_system.hpp"
 #include "resources/image.hpp"
 
 #define STB_IMAGE_IMPLEMENTATION
