+++ magnolia/src/resources/model_loader.cpp
@@ -3,8 +3,8 @@
 // this header on top
 
 #include "core/buffer.hpp"
-#include "core/file_system.hpp"
 #include "core/logger.hpp"
+#include "platform/file_system.hpp"
 #include "resources/model.hpp"
 
 namespace mag
