+++ magnolia/src/resources/shader_loader.cpp
@@ -5,8 +5,8 @@
 #include <vulkan/vulkan.hpp>
 
 #include "core/buffer.hpp"
-#include "core/file_system.hpp"
 #include "core/logger.hpp"
+#include "platform/file_system.hpp"
 #include "renderer/context.hpp"
 #include "renderer/shader.hpp"
 #include "spirv_reflect.h"
