+++ magnolia/src/scene/scene_serializer.cpp
@@ -1,9 +1,9 @@
 #include "scene/scene_serializer.hpp"
 
 #include "core/application.hpp"
-#include "core/file_system.hpp"
 #include "ecs/components.hpp"
 #include "ecs/ecs.hpp"
+#include "platform/file_system.hpp"
 #include "renderer/test_model.hpp"
 #include "resources/image.hpp"
 #include "scene/scene.hpp"
