+++ sprout_editor/src/editor_scene.cpp
@@ -3,10 +3,10 @@
 #include "camera_controller.hpp"
 #include "core/application.hpp"
 #include "core/assert.hpp"
-#include "core/file_system.hpp"
 #include "core/window.hpp"
 #include "ecs/components.hpp"
 #include "ecs/ecs.hpp"
+#include "platform/file_system.hpp"
 #include "threads/job_system.hpp"
 
 namespace sprout
