+++ sprout_editor/src/panels/viewport_panel.cpp
@@ -7,7 +7,6 @@
 #include "camera/camera.hpp"
 #include "core/application.hpp"
 #include "core/event.hpp"
-#include "core/file_system.hpp"
 #include "core/logger.hpp"
 #include "ecs/ecs.hpp"
 #include "editor.hpp"
@@ -15,6 +14,7 @@
 #include "icon_font_cpp/IconsFontAwesome6.h"
 #include "imgui/misc/cpp/imgui_stdlib.h"
 #include "math/generic.hpp"
+#include "platform/file_system.hpp"
 #include "renderer/renderer_image.hpp"
 #include "renderer/sampler.hpp"
 #include "renderer/shader.hpp"
